FIXES:
I fixed the issue where the “invalid barcode” would pop up if a Q was
typed.

Separated a MASTERcatalog.txt and the CheckedOut.txt file. Currently,
the program runs the same as if there were only one file, but
eventually I will want to NEVER have the MASTERcatalog.txt touched, but
it will be only a reference file and the CheckedOut.txt will be written
to so that the materials that are being checked out can be referenced
but the program. I am thinking that ONLY the barcodes will be stored
here because I was having a hard time trying to get all the info there.
It actually seems SO unnecessary to keep having alllll the material
repeated.